backend/requirements.txt
djangorestframework
mysqlclient
pandas